+++ EnhanceQoL/NewSettingsTable.lua
@@ -1,33 +1,7 @@
 local addonName, addon = ...
 
 addon.variables.NewVersionTableEQOL = {
-	-- 10.17 root categories
-	EQOL_UI = true,
-
-	-- 10.17 expandables
-	EQOL_Nameplates = true,
-
-	-- 10.17 nameplate markers and text controls
-	EQOL_nameplateEliteMarkerAnchor = true,
-	EQOL_nameplateEliteMarkerSize = true,
-	EQOL_nameplateEliteMarkers = true,
-	EQOL_nameplateFocusHealthbarTexture = true,
-	EQOL_nameplateFriendlyPlayerClassColorNames = true,
-	EQOL_nameplateFriendlyPlayerNamesOnly = true,
-	EQOL_nameplateHideFriendlyPlayerRealms = true,
-	EQOL_nameplateQuestMarkerAnchor = true,
-	EQOL_nameplateQuestMarkerSize = true,
-	EQOL_nameplateQuestMarkers = true,
-	EQOL_nameplateSlugOutline = true,
-	EQOL_nameplateTargetMarkerAtlas = true,
-	EQOL_nameplateTargetMarkerSize = true,
-	EQOL_nameplateTargetMarkers = true,
-	EQOL_nameplateTextFont = true,
-	EQOL_nameplateTextOutline = true,
-	EQOL_nameplateTextSize = true,
-
-	-- 10.18.0-beta2 tooltip ID modifier controls
-	EQOL_Tooltip = true,
-	EQOL_TooltipIDModifier = true,
-	EQOL_TooltipIDRequireModifier = true,
+	-- 10.19.0 combat logging
+	EQOL_GAMEPLAY = true,
+	EQOL_combatLogRaidCurrentExpansionOnly = true,
 }